fix : auth관련 타입 conatant폴더에서 type폴더로 옮김
