+++ src/types/auth.ts
@@ -1,0 +1,10 @@
+interface OauthLoginRequest {
+  code: string;
+}
+
+interface OauthLoginResponse {
+  token: string;
+  isNewMember: boolean;
+}
+
+export { OauthLoginRequest, OauthLoginResponse };
